+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+Breaking
+  * AZ-393: Remove support of Terraform 0.12
+
 Added
   * AZ-393: Monitoring Function for extra Azure metrics